Add an install script and derive CIDR from a netmask
The file layout has been vastly simplified because there was no
good reason to have it laid out as a replica of the target layout
when the install script knows where to put things.

You can pass in the netmask of the network as a dynamic attribute
from vSphere (taken from the Network protocol policy) so it makes
sense to have a "netmask" OVF property from which we derive the
CIDR value rather than having a CIDR property.